[add] routerの構築(1)
